+++ jetbrains-rider/resources/META-INF/ext-rider.xml
@@ -45,9 +45,7 @@
             <add-to-group group-id="EditorPopupMenu" anchor="last"/>
             <add-to-group group-id="EditorTabPopupMenu" anchor="last"/>
 
-            <separator/>
-            <action id="serverless.application.deploy.rider" class="software.aws.toolkits.jetbrains.services.lambda.actions.DeployServerlessApplicationAction"/>
-            <separator/>
+
         </group>
     </actions>
